feat(postgresql): Move to standalone plugin
