--- cli/internal/destinations/bigquery.go
@@ -1,1 +1,0 @@
-package destinations
