--- cli/internal/destinations/kafka.go
@@ -1,1 +1,0 @@
-package destinations
