+++ plugins/destination/postgresql/go.mod
@@ -1,3 +1,7 @@
+<<<<<<< HEAD
 module github.com/cloudquery/cloudquery/plugins/destination/postgresql
+=======
+module github.com/cloudquery/plugins/destinations/postgresql
+>>>>>>> a9ed9b1eb (feat(postgresql): Move to standalone plugin)
 
 go 1.19
